Move CC and TRC exceptions to a separate header file
